Renamed feature class and updated package
