+++ .vscode/launch.json
@@ -10,7 +10,7 @@
             "request": "launch",
             "classPaths": ["/Users/richard.somers@ibm.com/Documents/github/Galasa/framework/galasa-parent/galasa-boot/target/galasa-boot-0.5.0-SNAPSHOT.jar"],
             "mainClass": "dev.galasa.boot.Launcher",
-            "args": "--localmaven file:/Users/richard.somers@ibm.com/.m2/repository --obr mvn:dev.galasa/dev.galasa.uber.obr/0.5.0-SNAPSHOT/obr --obr mvn:dev.galasa/dev.galasa.simbank.obr/0.5.0-SNAPSHOT/obr --test dev.galasa.simbank.tests/test"
+            "args": "--localmaven file:/Users/richard.somers@ibm.com/.m2/repository --obr mvn:dev.galasa/dev.galasa.uber.obr/0.5.0-SNAPSHOT/obr --obr mvn:dev.galasa/dev.galasa.simbank.obr/0.5.0-SNAPSHOT/obr --test dev.galasa.simbank.tests/dev.galasa.translatedghrekin.SimbankGhrekin"
         },
         {
             "type": "java",
